63627692aa971 Web Portal Security Remediation Spectros: Web Portal Security Shipping	admin_shipment Site	configurations Site	content_block Site	page Site	search [N/A] Storage	browse Storage	file

+++ modules/site/default/configurations_mc.php
@@ -5,20 +5,27 @@
 	
     $siteConfiguration = new \Site\Configuration();
     if (!empty($_REQUEST['key'])) $siteConfiguration = new \Site\Configuration($_REQUEST['key']);
-    if (isset($_REQUEST['todo'])) {
-        switch ($_REQUEST['todo']) {
-            case 'drop':
-                $siteConfiguration->delete();
-                break;
+    if (isset($_REQUEST['todo']) && !empty($_REQUEST['todo'])) {
+        if (! $GLOBALS['_SESSION_']->verifyCSRFToken($_POST['csrfToken'])) {
+            $page->addError("Invalid Request");
+        } else {
+            switch ($_REQUEST['todo']) {
+                case 'drop':
+                    $siteConfiguration->delete();
+                    $page->success = 'Configuration Deleted';
+                    break;
 
-            case 'add':
-                $siteConfiguration->key = $_REQUEST['key'];
-                $siteConfiguration->set($_REQUEST['value']);
-                break;
-                
-            case 'update':
-                $siteConfiguration->set($_REQUEST['value']);
-                break;
+                case 'add':
+                    $siteConfiguration->key = $_REQUEST['key'];
+                    $siteConfiguration->set($_REQUEST['value']);
+                    $page->success = 'Configuration Added';
+                    break;
+                    
+                case 'update':
+                    $siteConfiguration->set($_REQUEST['value']);
+                    $page->success = 'Configuration Update';
+                    break;
+            }
         }
     }
